# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tariff Comparison APIs",
        "description": "Tariff Comparison APIs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "tready.aparna@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/",
            "description": "Tariff Comparison OpenApi Host"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "admin authentication"
                ],
                "summary": "Admin Login",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "user@test.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Valid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "example": 200
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number",
                                                            "example": 2
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "User"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "user@test.com"
                                                        },
                                                        "emailVerifiedAt": {
                                                            "type": "string",
                                                            "example": null
                                                        },
                                                        "updatedAt": {
                                                            "type": "string",
                                                            "example": "2022-06-28 06:06:17"
                                                        },
                                                        "createdAt": {
                                                            "type": "string",
                                                            "example": "2022-06-28 06:06:17"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "access_token": {
                                                    "properties": {
                                                        "token": {
                                                            "type": "string",
                                                            "example": "randomtokenasfhajskfhajf398rureuuhfdshk"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "Bearer"
                                                        },
                                                        "expires_in": {
                                                            "type": "number",
                                                            "example": 3600
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "example": 401
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "error"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Incorrect username or password!"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "admin authentication"
                ],
                "summary": "Admin Register",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "User",
                                    "email": "user@test.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "example": 200
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "User"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "user@test.com"
                                                        },
                                                        "roleId": {
                                                            "type": "integer",
                                                            "example": "1"
                                                        },
                                                        "emailVerifiedAt": {
                                                            "type": "string",
                                                            "example": null
                                                        },
                                                        "updatedAt": {
                                                            "type": "string",
                                                            "example": "2022-06-28 06:06:17"
                                                        },
                                                        "createdAt": {
                                                            "type": "string",
                                                            "example": "2022-06-28 06:06:17"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "access_token": {
                                                    "properties": {
                                                        "token": {
                                                            "type": "string",
                                                            "example": "randomtokenasfhajskfhajf398rureuuhfdshk"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "Bearer"
                                                        },
                                                        "expires_in": {
                                                            "type": "number",
                                                            "example": 3600
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "example": 422
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "error"
                                                },
                                                "message": {
                                                    "properties": {
                                                        "email": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "The email has already been taken."
                                                            },
                                                            "collectionFormat": "multi"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}